LootReceived event split into Npc & Player events
