+++ runelite-client/src/main/java/net/runelite/client/game/loot/events/NpcLootReceived.java
@@ -22,19 +22,30 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.game.loot;
+package net.runelite.client.game.loot.events;
 
-public class LootTypes
+import lombok.Data;
+import lombok.Getter;
+import net.runelite.api.NPC;
+import net.runelite.api.NPCComposition;
+import net.runelite.api.coords.WorldPoint;
+import java.util.Map;
+
+/**
+ * An event where you killed an {@link NPC} and received loot
+ */
+@Data
+public class NpcLootReceived
 {
-	public static final String UNKNOWN_EVENT = "UNKNOWN EVENT";
-	public static final String BARROWS = "BARROWS";
-	public static final String RAIDS = "RAIDS";
-	public static final String CHAMBERS_OF_XERIC = "RAIDS";
-	public static final String RAIDS_2 = "THEATRE OF BLOOD";
-	public static final String THEATRE_OF_BLOOD = "THEATRE OF BLOOD";
-	public static final String CLUE_SCROLL_EASY = "CLUE SCROLL EASY";
-	public static final String CLUE_SCROLL_MEDIUM = "CLUE SCROLL MEDIUM";
-	public static final String CLUE_SCROLL_HARD = "CLUE SCROLL HARD";
-	public static final String CLUE_SCROLL_ELITE = "CLUE SCROLL ELITE";
-	public static final String CLUE_SCROLL_MASTER = "CLUE SCROLL MASTER";
+	@Getter
+	private final int npcId;
+
+	@Getter
+	private final NPCComposition composition;
+
+	@Getter
+	private final WorldPoint location;
+
+	@Getter
+	private final Map<Integer, Integer> items;
 }
